refactor: move auth request/response DTOs to domain module and remove AuthMapper
Slice Pattern 표준(docs/ARCHITECTURE.md)에 맞춰 api/auth/controller 아래
있던 request/response DTO 9개를 domain/auth/dto/{request,response}로
이동. 레퍼런스 슬라이스 domain/comment가 이미 같은 위치에 Swagger
(@Schema) + Jakarta Validation 어노테이션을 포함한 DTO를 두고 있어
컨벤션이 이미 확립되어 있었음.

AuthMapper (MapStruct)는 제거. 기존 9개 메서드 중 6개는 이미 수동
default 구현이었고, 나머지 3개(toSignUpCommand, toSignUpResponse,
toKakaoLinkCommand)도 필드 이름이 1:1로 일치하는 단순 복사라 MapStruct
의존성이 실질 가치를 제공하지 못했음. 컨트롤러 두 곳(AuthController,
UserController)에서 record 생성자 직접 호출로 인라인.

이동한 DTO:
- request/: ActionTokenRequest, EmailRequest, PasswordResetConfirmRequest,
  SignInRequest, SignUpRequest, SocialAuthorizeRequest, SocialLoginRequest
- response/: SignUpResponse, SocialAuthorizeResponse

영향 범위:
- api: AuthController (mapper 필드/호출 제거, 인라인), UserController
  (동일), 테스트 2곳에서 AuthMapperImpl 제거 및 import 경로 업데이트
- 빈 디렉터리 api/auth/controller/{request,response} 삭제

테스트: 전 모듈 통과.
