+++ eventitta-domain/src/main/java/com/eventitta/domain/auth/dto/request/SocialLoginRequest.java
@@ -1,4 +1,4 @@
-package com.eventitta.api.auth.controller.request;
+package com.eventitta.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
