+++ eventitta-domain/src/main/java/com/eventitta/domain/auth/dto/response/SocialAuthorizeResponse.java
@@ -1,4 +1,4 @@
-package com.eventitta.api.auth.controller.response;
+package com.eventitta.domain.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
